update2

+++ utils/src/main/java/com/epam/utils/StringUtils.java
@@ -2,7 +2,7 @@ package com.epam.utils;
 
 public class StringUtils {
     public static boolean isPositiveNumber(String str) {
-         return isNumeric(str) && Integer.parseInt(str) > 0;//here magic will happen
+         return isNumeric(str) && Double.parseDouble(str) > 0;//here magic will happen
     }
     private static boolean isNumeric(String str) {
         return str.matches("-?\\d+(\\.\\d+)?");
